requirements.txt
streamlit
bycrpt
pandas
plotly
google-genai
